# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tinder-like API",
        "description": "Backend API untuk interaksi orang (Laravel 12)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/tinder": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "List recommended people",
                "operationId": "46e996b8295fd69da236ea91c95ab8f4",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/tinder/interact": {
            "post": {
                "tags": [
                    "Interaction"
                ],
                "summary": "Like or dislike a person",
                "operationId": "0928a1e48f536fd13081588a2fc2bf3c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "person_id",
                                    "type"
                                ],
                                "properties": {
                                    "person_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "like",
                                            "dislike"
                                        ],
                                        "example": "like"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Interaction created"
                    },
                    "422": {
                        "description": "Person already liked"
                    }
                }
            }
        },
        "/api/liked": {
            "get": {
                "tags": [
                    "Interaction"
                ],
                "summary": "List of liked people",
                "operationId": "57759dda76c377c3f90f9ddeaeab3cd5",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "People",
            "description": "People"
        },
        {
            "name": "Interaction",
            "description": "Interaction"
        }
    ]
}